services/ai-tutor/requirements.txt
fastapi
uvicorn
grpcio
grpcio-tools
protobuf<6.0.0
pydantic
pydantic-settings
